Added buffer detection in iqmesh::generate.

+++ src/iqm.h
@@ -4,26 +4,26 @@
 struct iqmheader
 {
 	char magic[16];
-	unsigned int version;
-	unsigned int filesize;
-	unsigned int flags;
-	unsigned int num_text, ofs_text;
-	unsigned int num_meshes, ofs_meshes;
-	unsigned int num_vertexarrays, num_vertexes, ofs_vertexarrays;
-	unsigned int num_triangles, ofs_triangles, ofs_adjacency;
-	unsigned int num_joints, ofs_joints;
-	unsigned int num_poses, ofs_poses;
-	unsigned int num_anims, ofs_anims;
-	unsigned int num_frames, num_framechannels, ofs_frames, ofs_bounds;
-	unsigned int num_comment, ofs_comment;
-	unsigned int num_extensions, ofs_extensions;
+	uint32_t version;
+	uint32_t filesize;
+	uint32_t flags;
+	uint32_t num_text, ofs_text;
+	uint32_t num_meshes, ofs_meshes;
+	uint32_t num_vertexarrays, num_vertexes, ofs_vertexarrays;
+	uint32_t num_triangles, ofs_triangles, ofs_adjacency;
+	uint32_t num_joints, ofs_joints;
+	uint32_t num_poses, ofs_poses;
+	uint32_t num_anims, ofs_anims;
+	uint32_t num_frames, num_framechannels, ofs_frames, ofs_bounds;
+	uint32_t num_comment, ofs_comment;
+	uint32_t num_extensions, ofs_extensions;
 };
 struct iqmmesh
 {
-	unsigned int name;
-	unsigned int material;
-	unsigned int first_vertex, num_vertexes;
-	unsigned int first_triangle, num_triangles;
+	uint32_t name;
+	uint32_t material;
+	uint32_t first_vertex, num_vertexes;
+	uint32_t first_triangle, num_triangles;
 };
 enum
 {
@@ -51,11 +51,11 @@ enum
 };
 struct iqmtriangle
 {
-	glm::detail::tvec3<unsigned int> verts;
+	glm::detail::tvec3<uint8_t> verts;
 };
 struct iqmjoint
 {
-	unsigned int name;
+	uint32_t name;
 	int parent;
 	glm::vec3 translate,scale;
 	glm::vec4 rotate;
@@ -63,16 +63,16 @@ struct iqmjoint
 struct iqmpose
 {
 	int parent;
-	unsigned int mask;
+	uint32_t mask;
 	float channeloffset[10];
 	float channelscale[10];
 };
 struct iqmanim
 {
-	unsigned int name;
-	unsigned int first_frame, num_frames;
+	uint32_t name;
+	uint32_t first_frame, num_frames;
 	float framerate;
-	unsigned int flags;
+	uint32_t flags;
 };
 enum
 {
@@ -80,11 +80,11 @@ enum
 };
 struct iqmvertexarray
 {
-	unsigned int type;
-	unsigned int flags;
-	unsigned int format;
-	unsigned int size;
-	unsigned int offset;
+	uint32_t type;
+	uint32_t flags;
+	uint32_t format;
+	uint32_t size;
+	uint32_t offset;
 };
 struct iqmbounds
 {
